микрофронт cardboard почти закончен, осталось только вынести оттуда редактирования профиля, аватары и окно загрузки.
